ttat

+++ P4/tinyFS.h
@@ -97,6 +97,6 @@ fdNode* add(fdNode* head, int data, uint8_t blockNbr);
 int removeNode(fdNode* head, int data);
 int containsFD(fdNode* head, int fd);
 int getBlockNbr(fdNode* head, int fd);
-int getFreeBlocks(int nbr, uint8_t index_free);
+int getFreeBlocks(int nbr, int index_free);
 
 #endif
